Move Linkify to ui-common and use kotlin conventions
